+++ odbc-api/src/cursor.rs
@@ -5,7 +5,7 @@ use crate::{
     error::ExtendResult,
     handles::{AsStatementRef, SqlResult, State, Statement, StatementRef},
     parameter::{VarBinarySliceMut, VarCharSliceMut},
-    sleep::{wait_for, Sleep},
+    sleep::{Sleep, poll_until_completed},
     Error, OutputParameter, ResultSetMetadata,
 };
 
@@ -567,7 +567,8 @@ where
     ) -> Result<Option<&B>, Error> {
         let mut stmt = self.cursor.as_stmt_ref();
         unsafe {
-            let result = wait_for(|| stmt.fetch(), &mut sleep).await;
+            let mut result = stmt.fetch();
+            poll_until_completed(&mut result, "SQLFetch", &mut stmt, &mut sleep).await?;
             let has_row = error_handling_for_fetch(result, stmt, error_for_truncation)?;
             Ok(has_row.then_some(&self.buffer))
         }