+++ odbc-api/src/handles/sql_result.rs
@@ -9,8 +9,11 @@ pub enum SqlResult<T> {
     Success(T),
     /// The function has been executed successfully. There have been warnings.
     SuccessWithInfo(T),
-    /// No more data is available
+    /// Meaning depends on the function emitting `NoData`.
     NoData,
+    /// Emmitted by execute in case delayed parameters have been bound and their input values are
+    /// now required.
+    NeedData,
     /// The function was started asynchronously and is still executing.
     StillExecuting,
     /// The function returned an error state. Check diagnostics.
@@ -49,15 +52,14 @@ impl<T> SqlResult<T> {
             SqlResult::Error { function } => SqlResult::Error { function },
             SqlResult::StillExecuting => SqlResult::StillExecuting,
             SqlResult::NoData => SqlResult::NoData,
+            SqlResult::NeedData => SqlResult::NeedData,
         }
     }
 
     pub fn unwrap(self) -> T {
         match self {
             SqlResult::Success(v) | SqlResult::SuccessWithInfo(v) => v,
-            SqlResult::Error { .. } => panic!("Unwrapping SqlResult::Error"),
-            SqlResult::NoData => panic!("Unwrapping SqlResult::NoData"),
-            SqlResult::StillExecuting => panic!("Unwrapping SqlResult::StillExecuting"),
+            _ => panic!("Invalid unwrapping of SqlResult"),
         }
     }
 }